backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
mysql-connector-python==8.0.26
python-dotenv==0.19.0
reportlab==3.6.0
qrcode==7.3
python-barcode==0.13.1
Pillow==8.3.2
PyJWT==2.1.0
